+++ src/fs.rs
@@ -1,7 +1,7 @@
 use crate::{
     bytes::Bytes,
     conn::Connection,
-    mount::{Fusermount, MountOptions},
+    mount::{unpriv::Fusermount, MountOptions},
     op::{self, Forget, Operation},
     reply,
     request::{FallbackBuf, RequestBuf, RequestHeader, SpliceBuf},
@@ -15,12 +15,13 @@ use rustix::{
     process::Pid,
 };
 use std::{
+    borrow::Cow,
     ffi::OsStr,
     future::Future,
     io,
     num::NonZeroUsize,
     panic,
-    path::PathBuf,
+    path::Path,
     sync::{
         atomic::{AtomicU64, Ordering},
         Arc, Weak,
@@ -380,11 +381,11 @@ pub struct Daemon {
 
 impl Daemon {
     pub async fn mount(
-        mountpoint: PathBuf,
+        mountpoint: impl Into<Cow<'static, Path>>,
         mountopts: MountOptions,
         mut config: KernelConfig,
     ) -> io::Result<Self> {
-        let (conn, fusermount) = crate::mount::mount(mountpoint, mountopts)?;
+        let (conn, fusermount) = crate::mount::mount_unpriv(mountpoint.into(), &mountopts)?;
 
         let mut session = Session::new();
         session.init(&conn, &mut config)?;
