+++ src/mount.rs
@@ -1,28 +1,14 @@
-use super::conn::Connection;
+pub mod unpriv;
+
 use bitflags::{bitflags, bitflags_match};
-use rustix::{
-    io::FdFlags,
-    net::{RecvAncillaryBuffer, RecvAncillaryMessage, RecvFlags},
-};
-use std::{
-    borrow::Cow,
-    fmt, io,
-    mem::{ManuallyDrop, MaybeUninit},
-    os::{
-        fd::OwnedFd,
-        unix::{net::UnixStream, prelude::*},
-    },
-    path::{Path, PathBuf},
-    process::{Child, Command},
-};
+use std::{borrow::Cow, fmt, path::Path};
+
+pub use unpriv::mount_unpriv;
 
 // refs:
 // * https://github.com/libfuse/libfuse/blob/fuse-3.10.5/lib/mount.c
 // * https://github.com/libfuse/libfuse/blob/fuse-3.10.5/util/fusermount.c
 // * https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/fs/fuse/inode.c?h=v6.15.9
-
-const FUSERMOUNT_PROG: &str = "/usr/bin/fusermount";
-const FUSE_COMMFD_ENV: &str = "_FUSE_COMMFD";
 
 bitflags! {
     #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
@@ -200,141 +186,6 @@ impl MountOptions {
             fusermount_path: None,
         }
     }
-}
-
-fn fusermount_path(opts: &MountOptions) -> &Path {
-    opts.fusermount_path
-        .as_deref()
-        .unwrap_or_else(|| Path::new(FUSERMOUNT_PROG))
-}
-
-#[derive(Debug)]
-pub struct Fusermount {
-    child: Option<PipedChild>,
-    mountpoint: PathBuf,
-    mountopts: MountOptions,
-}
-
-impl Fusermount {
-    pub fn unmount(mut self) -> io::Result<()> {
-        self.unmount_()
-    }
-
-    fn unmount_(&mut self) -> io::Result<()> {
-        if let Some(child) = self.child.take() {
-            // この場合、fusermount の終了にともない umount(2) が暗黙的に呼び出される。
-            // なので、fd受信用の UnixStream を閉じてバックグラウンドの fusermount を終了する。
-            child.wait()?;
-        } else {
-            // fusermount は fd を受信した直後に終了しているので、明示的に umount(2) を呼ぶ必要がある。
-            // 非特権プロセスなので `fusermount -u /path/to/mountpoint` を呼ぶことで間接的にアンマウントを行う
-            unmount(&self.mountpoint, &self.mountopts)?;
-        }
-        Ok(())
-    }
-}
-
-impl Drop for Fusermount {
-    fn drop(&mut self) {
-        let _ = self.unmount_();
-    }
-}
-
-#[derive(Debug)]
-struct PipedChild {
-    child: Child,
-    input: UnixStream,
-}
-
-impl PipedChild {
-    fn wait(mut self) -> io::Result<()> {
-        drop(self.input);
-        let _st = self.child.wait()?;
-        Ok(())
-    }
-}
-
-/// Acquire the connection to the FUSE kernel driver associated with the specified mountpoint.
-pub fn mount(mountpoint: PathBuf, mountopts: MountOptions) -> io::Result<(Connection, Fusermount)> {
-    tracing::debug!("Mount information:");
-    tracing::debug!("  mountpoint: {:?}", mountpoint);
-    tracing::debug!("  opts: {:?}", mountopts);
-
-    let mut fusermount = Command::new(fusermount_path(&mountopts));
-
-    let opts = mountopts.to_string();
-    if !opts.is_empty() {
-        fusermount.arg("-o").arg(opts);
-    }
-
-    fusermount.arg("--").arg(&mountpoint);
-
-    let (input, output) = UnixStream::pair()?;
-
-    fusermount.env(FUSE_COMMFD_ENV, output.as_raw_fd().to_string());
-
-    unsafe {
-        let output = ManuallyDrop::new(output);
-        fusermount.pre_exec(move || {
-            rustix::io::fcntl_setfd(&*output, FdFlags::empty()).map_err(Into::into)
-        });
-    }
-
-    let child = fusermount.spawn()?;
-
-    let fd = receive_fd(&input)?;
-
-    let mut child = Some(PipedChild { child, input });
-    if !mountopts.flags.contains(MountFlags::AUTO_UNMOUNT) {
-        // When auto_unmount is not specified, `fusermount` exits immediately
-        // after sending the file descriptor and thus we need to wait until
-        // the command is exited.
-        let child = child.take().unwrap();
-        child.wait()?;
-    }
-
-    Ok((
-        Connection::from(fd),
-        Fusermount {
-            child,
-            mountpoint,
-            mountopts,
-        },
-    ))
-}
-
-fn receive_fd(reader: &UnixStream) -> io::Result<OwnedFd> {
-    let mut buf = [0u8; 1];
-    let mut space = [MaybeUninit::uninit(); rustix::cmsg_space!(ScmRights(1))];
-    let mut cmsg_buffer = RecvAncillaryBuffer::new(&mut space);
-
-    let _msg = rustix::net::recvmsg(
-        reader,
-        &mut [io::IoSliceMut::new(&mut buf[..])],
-        &mut cmsg_buffer,
-        RecvFlags::empty(),
-    );
-
-    let fd = cmsg_buffer
-        .drain()
-        .flat_map(|msg| match msg {
-            RecvAncillaryMessage::ScmRights(mut fds) => fds.next(),
-            _ => None,
-        })
-        .next()
-        .ok_or_else(|| io::Error::new(io::ErrorKind::Other, "recv_fd"))?;
-
-    rustix::io::fcntl_setfd(&fd, FdFlags::CLOEXEC)?;
-
-    Ok(fd)
-}
-
-fn unmount(mountpoint: &Path, mountopts: &MountOptions) -> io::Result<()> {
-    let _st = Command::new(fusermount_path(mountopts))
-        .args(["-u", "-q", "-z", "--"])
-        .arg(mountpoint)
-        .status()?;
-    Ok(())
 }
 
 #[cfg(test)]
